Move constant from collection.js to a different file
